RedScraper/RedScraper/requirements.txt
aiofiles==0.6.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
async-generator==1.10
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
distlib==0.3.4
dominate==2.6.0
filelock==3.4.2
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
frozenlist==1.2.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
numpy==1.21.5
pandas==1.3.5
pipenv==2022.1.8
platformdirs==2.4.1
praw==7.5.0
prawcore==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.29
typing_extensions==4.0.1
update-checker==0.18.0
urllib3==1.26.7
virtualenv==20.13.0
virtualenv-clone==0.5.7
visitor==0.1.3
websocket-client==1.2.3
Werkzeug==2.0.2
WTForms==3.0.1
yarl==1.7.2
